% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrichPathway.R
\name{ALLEXTID.Reactome}
\alias{ALLEXTID.Reactome}
\title{ALLEXTID.Reactome}
\usage{
\method{ALLEXTID}{Reactome}(organism)
}
\arguments{
\item{organism}{organism}
}
\description{
ALLEXTID.Reactome
}

