% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DOSE.R
\name{dotplot}
\alias{dotplot}
\title{dotplot}
\usage{
dotplot(object, ...)
}
\arguments{
\item{object}{an instance of enrichResult}

\item{...}{additional parameter}
}
\value{
ggplot object
}
\description{
dot plot of enrichResult
}
\author{
Guangchuang Yu
}

