% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichPathway.R
\name{enrichPathway}
\alias{enrichPathway}
\title{Pathway Enrichment Analysis of a gene set.
Given a vector of genes, this function will return the enriched pathways
with FDR control.}
\usage{
enrichPathway(
  gene,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe = NULL,
  weight = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  organism = "human",
  readable = FALSE
)
}
\arguments{
\item{gene}{a vector of entrez gene id.}

\item{pvalueCutoff}{Cutoff value of pvalue.}

\item{pAdjustMethod}{one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"}

\item{universe}{background genes}

\item{weight}{A named numeric vector of weights for background genes. If provided, Weighted ORA will be performed using Wallenius' noncentral hypergeometric distribution.}

\item{minGSSize}{minimal size of genes annotated by Ontology term for testing.}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{qvalueCutoff}{Cutoff value of qvalue}

\item{organism}{one of "human", "rat", "mouse", "celegans", "yeast", "zebrafish", "fly".}

\item{readable}{whether mapping gene ID to gene Name}
}
\value{
A \code{enrichResult} instance.
}
\description{
Pathway Enrichment Analysis of a gene set.
Given a vector of genes, this function will return the enriched pathways
with FDR control.
}
\examples{
\dontrun{
	gene <- c("11171", "8243", "112464", "2194",
			"9318", "79026", "1654", "65003",
			"6240", "3476", "6238", "3836",
			"4176", "1017", "249")
	yy = enrichPathway(gene, pvalueCutoff=0.05)
	head(summary(yy))
}
}
\seealso{
\code{\link[enrichit]{enrichResult-class}}
}
\author{
Guangchuang Yu \url{http://ygc.name}
}
\keyword{manip}
