\docType{class}
\name{enrichPathwayResult-class}
\alias{enrichPathwayResult-class}
\alias{plot,enrichPathwayResult-method}
\alias{show,enrichPathwayResult-method}
\alias{summary,enrichPathwayResult-method}
\title{Class "enrichPathwayResult"}
\description{
  This class represents the result of Pathway enrichment
  analysis.
}
\section{Methods}{

  In the code snippets below, \code{x} is a \code{enrichPathwayResult} object.

\describe{
  \item{show}{\code{show(x)} show analysis summary.}
  \item{summary}{\code{summary(x)} return enrichment result in a data frame.}
  \item{plot}{\code{plot(x)} visualize enrichment result.}
}
See \code{?\link{show}}, \code{?\link{summary}}, and \code{?\link{plot}} for details of operations on enrichPathwayResult objects.
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\seealso{
  \code{\link{enrichPathway}}
}
\keyword{classes}

