% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrichPathway.R
\name{EXTID2TERMID.Reactome}
\alias{EXTID2TERMID.Reactome}
\title{EXTID2TERMID.Reactome}
\usage{
\method{EXTID2TERMID}{Reactome}(gene, organism)
}
\arguments{
\item{gene}{gene}

\item{organism}{organism}
}
\description{
EXTID2TERMID.Reactome
}

