% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getGeneSet.Reactome}
\alias{getGeneSet.Reactome}
\title{getGeneSet.Reactome}
\usage{
\method{getGeneSet}{Reactome}(setType = "Reactome", organism)
}
\arguments{
\item{setType}{gene set type}

\item{organism}{organism}
}
\description{
getGeneSet.Reactome
}

