% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gson.R
\name{gson_Reactome}
\alias{gson_Reactome}
\title{gson_Reactome}
\usage{
gson_Reactome(organism = "human")
}
\arguments{
\item{organism}{one of "human", "rat", "mouse", "celegans", "yeast", "zebrafish", "fly".}
}
\value{
a 'GSON' object
}
\description{
download the latest version of Reactome and stored in a 'GSON' object
}
\examples{
\dontrun{
rec_gson <- gson_Reactome("human")
}

}
