\name{pathID2Name}
\alias{pathID2Name}
\title{pathID2Name}
\usage{
  pathID2Name(pathID)
}
\arguments{
  \item{pathID}{query Path ID}
}
\value{
  Pathway Name
}
\description{
  Mapping Pathway ID to Pathway Name
}
\examples{
pathID2Name(c("1221632","75983"))
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}
\keyword{mainip}

