\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,enrichPathwayResult,ANY-method}
\title{plot method}
\arguments{
  \item{x}{A \code{enrichPathwayResult} instance.}

  \item{...}{Additional argument list}
}
\value{
  plot
}
\description{
  plot method for \code{enrichPathwayResult} instance
}
\author{
  Guangchuang Yu \url{http://ygc.name}
}

