% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSE.R, R/enrichplot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{geneID}
\alias{geneInCategory}
\alias{cnetplot}
\alias{dotplot}
\alias{emapplot}
\alias{gseaplot}
\alias{heatplot}
\alias{ridgeplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{enrichit}{\code{\link[enrichit]{geneID}}, \code{\link[enrichit]{geneInCategory}}}

  \item{enrichplot}{\code{\link[enrichplot:reexports]{cnetplot}}, \code{\link[enrichplot]{dotplot}}, \code{\link[enrichplot]{emapplot}}, \code{\link[enrichplot]{gseaplot}}, \code{\link[enrichplot]{heatplot}}, \code{\link[enrichplot]{ridgeplot}}}
}}

