% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrichPathway.R
\name{TERM2NAME.Reactome}
\alias{TERM2NAME.Reactome}
\title{TERM2NAME.Reactome}
\usage{
\method{TERM2NAME}{Reactome}(term, organism)
}
\arguments{
\item{term}{reactome pathway term}

\item{organism}{organism}
}
\description{
TERM2NAME.Reactome
}

