% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrichPathway.R
\name{TERMID2EXTID.Reactome}
\alias{TERMID2EXTID.Reactome}
\title{TERMID2EXTID.Reactome}
\usage{
\method{TERMID2EXTID}{Reactome}(term, organism)
}
\arguments{
\item{term}{reactome pathway term}

\item{organism}{organism}
}
\description{
TERMID2EXTID.Reactome
}

