% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewPathway.R
\name{viewPathway}
\alias{viewPathway}
\title{viewPathway}
\usage{
viewPathway(pathName, organism = "human", readable = TRUE,
  foldChange = NULL, layout = "kk", ...)
}
\arguments{
\item{pathName}{pathway Name}

\item{organism}{supported organism}

\item{readable}{logical}

\item{foldChange}{fold change}

\item{layout}{graph layout}

\item{...}{additional parameters}
}
\value{
plot
}
\description{
view reactome pathway
}
\details{
plotting reactome pathway
}
\author{
Yu Guangchuang
}
