% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseAnalyzer.R
\name{gsePathway}
\alias{gsePathway}
\title{gsePathway}
\usage{
gsePathway(
  geneList,
  organism = "human",
  exponent = 1,
  minGSSize = 10,
  maxGSSize = 500,
  eps = 1e-10,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  verbose = TRUE,
  seed = FALSE,
  by = "fgsea",
  ...
)
}
\arguments{
\item{geneList}{order ranked geneList}

\item{organism}{organism}

\item{exponent}{weight of each step}

\item{minGSSize}{minimal size of each geneSet for analyzing}

\item{maxGSSize}{maximal size of each geneSet for analyzing}

\item{eps}{This parameter sets the boundary for calculating the p value.}

\item{pvalueCutoff}{pvalue Cutoff}

\item{pAdjustMethod}{pvalue adjustment method}

\item{verbose}{print message or not}

\item{seed}{logical}

\item{by}{one of 'fgsea' or 'DOSE'}

\item{...}{other parameter}
}
\value{
gseaResult object
}
\description{
Gene Set Enrichment Analysis of Reactome Pathway
}
\examples{
\dontrun{
library(ReactomePA)
data(geneList, package="DOSE")
y <- gsePathway(geneList, 
                organism = "human",
                pvalueCutoff = 0.2,
                pAdjustMethod = "BH", 
                verbose = FALSE)
head(y)
y2 <- gsePathway(geneList, 
                 organism = rec_gson,
                 pvalueCutoff = 0.2,
                 pAdjustMethod = "BH", 
                 verbose = FALSE)
head(y2)
}
}
\author{
Yu Guangchuang
}
