% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReactomePA-package.R
\docType{package}
\name{ReactomePA-package}
\alias{ReactomePA}
\alias{ReactomePA-package}
\title{ReactomePA: Reactome Pathway Analysis}
\description{
This package provides functions for pathway analysis based on REACTOME pathway database. It implements enrichment analysis, gene set enrichment analysis and several functions for visualization. This package is not affiliated with the Reactome team.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/biomedical-knowledge-mining-book/}
  \item Report bugs at \url{https://github.com/GuangchuangYu/ReactomePA/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

Other contributors:
\itemize{
  \item Vladislav Petyuk \email{petyuk@gmail.com} [contributor]
}

}
\keyword{internal}
