\docType{package}
\name{ReactomePA-package}
\alias{ReactomePA}
\alias{ReactomePA-package}
\title{Reactome Pathway Analysis}
\description{
  This package is designed for reactome pathway analysis.
}
\details{
  \tabular{ll}{ Package: \tab ReactomePA\cr Type: \tab
  Package\cr Version: \tab 0.2.1\cr Date: \tab
  02-09-2012\cr biocViews:\tab Bioinformatics, Pathway,
  Visulization\cr Depends:\tab AnnotationDbi, org.Hs.eg.db,
  igraph, plyr, methods, stats, qvalue, reactome.db\cr
  Suggests:\tab GOSemSim, DOSE, clusterProfiler\cr License:
  \tab GPL-2\cr }
}
\author{
  Guangchuang Yu \email{guangchuangyu@gmail.com}

  Maintainer: Guangchuang Yu
  \email{guangchuangyu@gmail.com}
}
\seealso{
  \linkS4class{enrichResult}
}
\keyword{package}

