% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{geneID}
\alias{geneInCategory}
\alias{cnetplot}
\alias{dotplot}
\alias{emapplot}
\alias{gseaplot}
\alias{heatplot}
\alias{ridgeplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{enrichit}{\code{\link[enrichit:geneID]{geneID()}}, \code{\link[enrichit:geneInCategory]{geneInCategory()}}}

  \item{enrichplot}{\code{\link[enrichplot:cnetplot]{cnetplot}}, \code{\link[enrichplot:dotplot]{dotplot()}}, \code{\link[enrichplot:emapplot]{emapplot()}}, \code{\link[enrichplot:gseaplot]{gseaplot()}}, \code{\link[enrichplot:heatplot]{heatplot()}}, \code{\link[enrichplot:ridgeplot]{ridgeplot()}}}
}}

