% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TERM2NAME.Reactome}
\alias{TERM2NAME.Reactome}
\title{TERM2NAME.Reactome}
\usage{
\method{TERM2NAME}{Reactome}(term, organism)
}
\arguments{
\item{term}{reactome pathway term}

\item{organism}{organism}
}
\description{
TERM2NAME.Reactome
}

