% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TERMID2EXTID.Reactome}
\alias{TERMID2EXTID.Reactome}
\title{TERMID2EXTID.Reactome}
\usage{
\method{TERMID2EXTID}{Reactome}(term, organism)
}
\arguments{
\item{term}{reactome pathway term}

\item{organism}{organism}
}
\description{
TERMID2EXTID.Reactome
}

