% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewPathway.R
\name{viewPathway}
\alias{viewPathway}
\title{viewPathway}
\usage{
viewPathway(
  pathName,
  organism = "human",
  readable = TRUE,
  foldChange = NULL,
  keyType = "ENTREZID",
  layout = "kk"
)
}
\arguments{
\item{pathName}{pathway Name}

\item{organism}{supported organism}

\item{readable}{logical}

\item{foldChange}{fold change}

\item{keyType}{keyType of gene ID (i.e. names of foldChange, if available)}

\item{layout}{graph layout}
}
\value{
plot
}
\description{
view reactome pathway
}
\details{
plotting reactome pathway
}
\author{
Yu Guangchuang
}
