\name{viewPathway}
\alias{viewPathway}
\title{viewPathway}
\usage{
viewPathway(pathName, organism = "human", readable = TRUE,
  foldChange = NULL, ...)
}
\arguments{
  \item{pathName}{pathway Name}

  \item{organism}{supported organism}

  \item{readable}{logical}

  \item{foldChange}{fold change}

  \item{...}{additional parameter}
}
\value{
plot
}
\description{
view reactome pathway
}
\details{
plotting reactome pathway
}
\author{
Yu Guangchuang
}

